ScrollView 삭제 된 DetailView 리팩토링
