+++ iOS/BaeminMap/BaeminMap/Extension.swift
@@ -34,3 +34,12 @@ extension Double {
 extension UIColor {
     static let pointColor = UIColor(red: 42/255, green: 193/255, blue: 188/255, alpha: 1)
 }
+
+extension UIScrollView {
+    func scrollToSection(y: CGFloat) {
+        let bottomOffset = CGPoint(x: 0, y: contentOffset.y + y)
+        if(bottomOffset.y > 0) {
+            setContentOffset(bottomOffset, animated: true)
+        }
+    }
+}
