--- iOS/BaeminMap/BaeminMap/TestTableViewCell.swift
@@ -1,32 +1,0 @@
-//
-//  TestTableViewCell.swift
-//  BaeminMap
-//
-//  Created by HannaJeon on 2017. 8. 15..
-//  Copyright © 2017년 HannaJeon. All rights reserved.
-//
-
-import UIKit
-
-class TestTableViewCell: UITableViewCell {
-
-//    @IBOutlet weak var collectionView: UICollectionView!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        
-    }
-    
-
-}
-//
-//extension TestTableViewCell: UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
-//    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-//        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "collectionViewCell", for: indexPath)
-//        return cell
-//    }
-//
-//    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-//        return 6
-//    }
-//
